grpc-js: Refactor in preparation for retries
