+++ packages/grpc-js/src/filter.ts
@@ -15,12 +15,14 @@
  *
  */
 
-import { Call, StatusObject, WriteObject } from './call-stream';
+import { StatusObject, WriteObject } from './call-interface';
 import { Metadata } from './metadata';
 
 /**
  * Filter classes represent related per-call logic and state that is primarily
- * used to modify incoming and outgoing data
+ * used to modify incoming and outgoing data. All async filters can be
+ * rejected. The rejection error must be a StatusObject, and a rejection will
+ * cause the call to end with that status.
  */
 export interface Filter {
   sendMetadata(metadata: Promise<Metadata>): Promise<Metadata>;
@@ -32,8 +34,6 @@ export interface Filter {
   receiveMessage(message: Promise<Buffer>): Promise<Buffer>;
 
   receiveTrailers(status: StatusObject): StatusObject;
-
-  refresh(): void;
 }
 
 export abstract class BaseFilter implements Filter {
@@ -56,10 +56,8 @@ export abstract class BaseFilter implements Filter {
   receiveTrailers(status: StatusObject): StatusObject {
     return status;
   }
-
-  refresh(): void {}
 }
 
 export interface FilterFactory<T extends Filter> {
-  createFilter(callStream: Call): T;
+  createFilter(): T;
 }
