+++ packages/grpc-js/src/load-balancer-outlier-detection.ts
@@ -15,8 +15,7 @@
  *
  */
 
-import { ChannelOptions, connectivityState, StatusObject } from ".";
-import { Call } from "./call-stream";
+import { ChannelOptions } from "./channel-options";
 import { ConnectivityState } from "./connectivity-state";
 import { Status } from "./constants";
 import { durationToMs, isDuration, msToDuration } from "./duration";
@@ -311,28 +310,6 @@ interface MapEntry {
   subchannelWrappers: OutlierDetectionSubchannelWrapper[];
 }
 
-class OutlierDetectionCounterFilter extends BaseFilter implements Filter {
-  constructor(private callCounter: CallCounter) {
-    super();
-  }
-  receiveTrailers(status: StatusObject): StatusObject {
-    if (status.code === Status.OK) {
-      this.callCounter.addSuccess();
-    } else {
-      this.callCounter.addFailure();
-    }
-    return status;
-  }
-}
-
-class OutlierDetectionCounterFilterFactory implements FilterFactory<OutlierDetectionCounterFilter> {
-  constructor(private callCounter: CallCounter) {}
-  createFilter(callStream: Call): OutlierDetectionCounterFilter {
-    return new OutlierDetectionCounterFilter(this.callCounter);
-  }
-
-}
-
 class OutlierDetectionPicker implements Picker {
   constructor(private wrappedPicker: Picker) {}
   pick(pickArgs: PickArgs): PickResult {
@@ -344,7 +321,14 @@ class OutlierDetectionPicker implements Picker {
         return {
           ...wrappedPick,
           subchannel: subchannelWrapper.getWrappedSubchannel(),
-          extraFilterFactories: [...wrappedPick.extraFilterFactories, new OutlierDetectionCounterFilterFactory(mapEntry.counter)]
+          onCallEnded: statusCode => {
+            if (statusCode === Status.OK) {
+              mapEntry.counter.addSuccess();
+            } else {
+              mapEntry.counter.addFailure();
+            }
+            wrappedPick.onCallEnded?.(statusCode);
+          }
         };
       } else {
         return wrappedPick;
